+++ CHANGELOG.md
@@ -8,6 +8,7 @@ Released on 2017-10-29
 
 ### Bug-fixes
 
+- [x] [#685 - JSON scheme accepts allow without default](https://github.com/scriptPilot/app-framework/issues/685)
 - [x] [#689 - Signout button does not sign out properly if complete app is is login-protected](https://github.com/scriptPilot/app-framework/issues/689) ([read the docs](https://github.com/scriptPilot/app-framework/blob/master/docs/firebase.md))
 - [x] [#693 - Login popup component error](https://github.com/scriptPilot/app-framework/issues/693)